Remove usages of .classType
They are no longer necessary with @Resolver or resolveType function.

I updated DatabaseService to use AbstractClassType instead of Type so it works for those functions as well.

Creating Union types had to have their type array be cast to any because they still want non-abstract classes.
But this was only two places, so it's worth it.