Fix ValidationException handling of nested constraint violations
Sometimes ValidationErrors don't have any constraint violations
but they have children that do. These were not being handled.
Now we flatten the list down and ignore parents that don't have
their own constraint violations.

Since we are using the target name to identify the nested object
there's no need to prefix the parent.
This is inverse to what is exposed to the API which identifies by
a nested path structure.

This was also missed because the shape of nestjs' ValidationError
doesn't have nullability correct. Switch to the interface from
class-validator produced the error correctly.