Fix event handler registration confusion from shared names
Due to reflection event handlers are identified by their
class names. This means that if there are multiple handlers
named the same, all of their listened-events will be shared
amongst them. This is most likely a copy/paste error and not
a real problem so we'll just check for duplicates and throw
an exception.