Update root db object setup & indexing to wait for database connection
This allows the webserver to start and health checks to pass even if
the database is failing.
This reduces strain on ops that have nothing to do with the node API.

For tests we still wait for the database connection so that this process
can still throw an error if we've done something wrong.
For production we just log the failure.