Refactor db transactions
Previously we were wrapping the neo4j transaction object
so it could work nicely with cypher query builder. However,
this required that the transaction object would be explicitly
passed around everywhere. This didn't fit well with the way we
were abstracting methods, since the db was never passed in but
rather injected into the class.

The bigger problem is that we were not using the builtin retry
logic. neo4j driver will retry logic if it encounters a transient
error, like connection problems & constraint changes.
It only uses this if the execution is inside of the
`session.readTransaction/writeTransaction` callbacks.

So solution. We are now exclusively using those to read/write transaction
wrappers. We are storing the transaction object in a native `AsyncLocalStorage`.
This allows us to fetch the current transaction regardless of which async
context we are in (there can be multiple transactions at once for multiple requests).
This allows us to patch the session runner to use the current transaction runner instead.
This means that all the current db code will "just work" on the transaction when executed
inside of a `runInTransaction` call.
This also means that `runInTransaction` calls can be nested without creating multiple transactions.
Only caveat is it requires Node 12.17+ or 14+, which shouldn't be a problem.