Ensure budget records are unique when creating
This prevents a timing problem between when a user's permissions
are created for a newly created budget record and another call which modifies budget records.
In this case the user will not have permission to see the new record and thus
our sync handler will create another one. This prevents that.

It could be deemed unnecessary once the sync handler doesn't depend on user permissions.