Created & tested DateInterval which is a range for CalendarDates
This is needed because we consider CORD date ranges to be inclusive of end dates.
i.e March 1st to March 3rd is 3 days long and covers the entire 3rd day.

This also has the start & end properties as CalendarDate instances
which allows them to automatically be transformed to Neo4j Dates