Change uploadPeriodicReport mutation to return the report.
Defined files are null to the API until the first version is uploaded.
This means PeriodicReport.reportFile changes when first version is uploaded.
Changing this mutation to return the report allows the UI to automatically
update the cached report's file instead of us having to make that connection manually.

It's also more consistent with other defined files & no extra db queries are needed.
So it's a win/win/win.