Validate parent node while creating directories/versions without attempting to hydrate parent
This fails on creating file versions of defined files, since it's possible defined files have
no existing file versions which means they will fail hydration ("not exist"). Normally this is
what we want while reading, but in this case we don't care because we are trying to create the
first version.