Convert session hydration from Pipe to Interceptor
This is necessary for all the queries that will reference a Loader and not a Session decorator.
Since the session object is always created, instead of just when a Session decroator is used,
the Loader will always have access to it.

I also moved the session to be stored on the gql context directly instead of in the request.
This further decouples us from express.

Note that it does not enforce that a user is logged in without using the LoggedInSession decorator.