Allow lazily references loader class to fix import execution
Also added runtime check to Loader() decorator to check for this case

This will throw a clear error at server start instead of a random one when a request for resolving a field fails.
Also allow type to be referenced lazily via function. Deferring like this allows node to get class from module
later in execution after circular references have been resolved.

This is the same thing that nestjs/graphql decorators do & for the same reason.