Refactor O(n) create periodic reports to O(1) merge reports
This creates all reports requested at once, this should help
performance since it needs to be done 3 times for each report type.
Financial reports for 2 years is 48 reports, so 48 queries down to 1.

The bigger issue this fixes is reports being duplicated.
A report could be pre-existing for a handful of reasons:
- Had a file attached, so it was never deleted
- Date range was changed before financial reporting period was set,
  When the period is set, the previous range doesn't account for all
  former changes.
- Transient errors?