Change DbLabel() resolution for class names
Previously the idea was that the decorator could override the whole parent tree. This wasn't useful.
It also had the mistake of looking for this decorator from any parent of the class, instead of just itself.

Now the decorator can only override the db label for its own class, and the parents are traversed regardless.