Don't fail request if there's no tracing segment
I tried to set up the middleware, so that requests always have a (root) segment.
This way the tracing functions could operate normally and if tracing was not
enabled for the request then the trace data would just be discarded.

This line however seems to be failing while debugging locally
https://github.com/SeedCompany/cord-api-v3/blob/40fa9424c8a4a78f87492550a4c61e129bfee4e0/src/core/tracing/xray.middleware.ts#L63-L63
Ideally this shouldn't happen. I think in the end though it's better to ignore
and move than fail. AWS still runs correctly, so that's really all that matters.