Replace changeset editability validation implementation
The previous code was guessing where a changeset argument would be,
and was defined by the consumer's operation. So easily defeated.

The new implementation validates based on what our API schema defines.
As long as input object uses the ChangesetIdField decorator it will be
validated in mutations regardless of key or depth.

This new implementation also uses the new interface property instead of
the ProjectChangeRequest concrete.