Add ChangesetArg
This was a previous iteration of the former commit, that I'm not ready to part ways with.
It could be argued that having changeset id as a root arg on each operation makes more
sense than what we currently do with input objects. This would be a breaking change though,
hence the other implementation in the former commit.
The former doesn't work when defining args directly, where as this does.
Although it would work if it was defined as an ArgsType object, which is defined in the schema
the same as this way.
I'll probably drop this code later, but I wanted it versioned just in case.