Refactor e2e gql client to fully decouple from src code
Now we make a POST request to the graphql endpoint,
which is exactly how any consumer would do it.
This saves us lots of business logic that we were doing before.

Previously we were jumping halfway into the server execution skipping some src code.
We in turn had to reimplement that code here (tracing, etc) and fake requests.

Also, because of this error handling was unique and confusing.
Now we to assert errors from the JSON response, which is better for e2e tests.