Change `Variable` to print `value` instead of `name`
The `name` is the bound parameter name, prefixed with `$`.
It is actually mutable to keep it unique in the context of the whole query.
This should be allowed & ignored, as variables are stripped out instead of
being turned into bound parameters.
The `value` is always the raw untouched cypher expression which is what we want.

With this change, we don't need to work around CQB's naming logic,
which simplifies the implementation as well.