Fix schema check not including what's already in base ref
Only specify ref for base case and use default ref for head case.
The default ref for pull requests actually creates a merge commit with
the base ref.
This is what we want here as the "diff" should be based on new changes
going into base, and excluding commits already in base.