Require filenames on defined file version upload (soft breaking)
This avoids an error where a name wasn't specified on the first defined file upload.
This was because we'd try to fetch the file which would be invalid,
because there weren't any versions.

In practice (in UI), the name is always given.
So while this is technically a breaking change,
it should be fine to roll out whenever.