Refactor url signing
The main need is to allow signing options to be passed through.
We want to the expiration duration to change based on different situations now.
This is better because the FileBuckets were just supposed to be the persistence layer,
so less business logic there is better.

I also took the opportunity to refactor to use aws-sdk v3 classes.
Doing it this way with generics is more elegant,
as we don't need to explicitly reference actions and their input options in bucket abstraction.