Change timezone data caching strategy
Previously, we only cached the data for 5 minutes.
This data actually only changes around once a month.

Instead, we will grab it at build and save it to the image.
This way it will be automatically fresh as of build time, and
we won't rely on GitHub in production.

For dev mode, we will just grab it anytime it doesn't exist
on disk.