Change how ConfigService is created to fix environment dependency
This has only been working due to the old down-level transpilation.
In short, modern ECMA calls class property initializers before calling
the constructor.
This means that the environment service will not be available.

The new process allows a similar syntax, but the env service injection
is done differently, so it's available for class property initializers.