Refactor how we adjust the neo4j retry logic
The previous logic dug into the neo4j lib and monkey
patched the "should retry" check to look at our "previous" exceptions.

This is invasive and doesn't work with ESM where module exports are immutable.

Instead, we can change our transaction handling logic.
Now when an error is thrown,
before we hand it to neo4j's TransactionExecutor to check for "should retry",
we look for the first previous exception that's a Neo4jError.
We give that error to TransactionExecutor, and associate the original in a map.
Finally, after TransactionExecutor deems the error is not retryable,
we unwrap it from our previously saved mapping.

This is much better and assumes less about the Neo4j driver's inner workings.