Replace cypher CLI syntax highlighting library
The main reason for this is that the old library includes the TS DOM lib.
This added those global types to our codebase, and we have no
way to override that.

In general, though, this new library is much better at calling the syntax.
Though I did take some time to tweak it.

It's worth noting that it uses chalk which requires ESM.
So, wouldn't have been easy to use this before now.