Skip optimizing union conditions while they are on the execution hot path
Logically, this makes no change, so I'm guessing this is better.
It's probably faster to just check `A & B`,
rather than do all the work to create a new condition `AB`
and then check that.
Since this can be called for every property on every check,
we need to be sensitive to that.

If it happens once during boot up
or while needing a more polished output
when viewing as a human, that's a different story.