Change how Postable asks for __typename
Stop requiring implementations to declare the __typename field.
This apparently was tricky for TS as it was missed in the Project shapes.
This string can also be gotten from GraphQL or ResourceLoader,
so it's not always necessary to specify manually.

Instead, we'll change types to only ask for __typename
when the PostableRef is an actual Postable object.