Refactor ffprobe to use signed url instead of piped object stream
I think this ends up being better in every way.

NodeJS doesn't have to worry about piping streams,
and the hangs that were occurring there.

Additionally, even when the previous process worked, we got a detected duration of 0.
This is some different with ffprobe.
Input streams would give this, but files/urls are detected correctly.

I upgraded the migration as well to re-detect any existing media
that has a duration of 0.