Fix bad assumptions about Cypher MERGE with media saving
How it actually works is if the complete pattern doesn't exist,
then the complete pattern is created.
Which means if the media didn't exist, it would create a FileVersion node.
We never want to create a FileVersion node here,
so pull that out to a match
and only merge if we have a FileVersion given.