Move "root" directory creation logic to FileService
The previous logic stopped working because we started
assuming with hydrate that a FileNode would have an attachment point,
aka not be orphaned.

With the Project root directory previous process, it would be "orphaned"
at first until the next query ran relating the node to the project.

Now it's just created all at once, which is better anyways.