Only enable impersonation if logged in
Situation is: caller calls session query with impersonation while not logged in.
Previously we threw an error that the impersonation was unauthorized.
This prevented the caller from ever establishing a session to login with.

Now we will just ignore their impersonation input if they are not logged in.
This allows a session to be established and login to be preformed.
After that point any operations will continue to error out if they are not authorized.