Switch generated query builder's target `mts` -> `ts` to fix usage with jest
Originally switched to support ESM.
Jest struggles with the imports though, and I could not find a configuration that worked.
The generated files produced with the `ts` target are fine for our use,
aside from the import/export from 'edgedb/*'.
These need the file extension to work with ESM.
Example:
- import { Cardinality } from "edgedb/dist/reflection/index";
+ import { Cardinality } from "edgedb/dist/reflection/index.js";