Adjust enforcing project context during project insertion triggers
This is a bit better as it doesn't assume anything about the project context.
Only that the given project & projectContext combo match.

This fixes project insertion, which now has a budget inserted with a trigger.
And that budget / "project child" was trying to check that it was given
the right project/project-context combo.
But the project doesn't yet have itself within its context yet,
because that self-reference has to in a separate query after the insert.