Handle registering codecs whose IDs are not statically known
This allows custom scalars, that are not base scalars, to be registered.
This introspects the types from the connected db instance to determine IDs.

Since this requires a DB connection,
I put some work into doing this gracefully.
We'll try to connect and register immediately,
if there is a connection established quickly, all is good.
If not, then we'll register statically known scalar codecs immediately,
and then keep trying to connect & register in the background.