Make ResourcesHost sync
Previously we needed to import the legacy resource map from another file.
Now that's removed, so this no longer needs to be async.

Obviously, this has huge implications.

An argument could be made to keep this async
to allow for future functionality that needs to be async.
However, I can't currently think of what this would be.

Making this sync allows for more places to use the current functionality.
Currently trying to use this in ExceptionNormalizer, which is sync.
I could make that async instead.