New decoupled way to associated EdgeDB types to our Resources
This was causing problems with TS.
TS would follow/compute these DB types for inheritance and our Relations object,
which was extra work that we didn't need.
It also caused problems with inheritance, which we tried to solve with `abstractType()`

This new way allows us to map from Resources to EdgeDB types, without
TS needing to consider those DB types when evaluating our Resource types.