Fix ChangesOf mapping of DefinedFiles
`FileId` used to have a unique brand that TS could discriminate on.
I removed this in a50ae3c to be forward compatible with LinkTo<File> approach.
So that removal broke this code.

Unfortunately, it was immediately hidden by my next commit 2b9f01e
which was also bugged.
And I think, caused all change props to be `unknown`
which TS was somehow ok with.

Really the only logical change here is that:
ID (== ID<any>) are not considered File IDs (ID<File>).
Which takes an extra step because `any` is assignable to `File`,
so it has to be explicitly blocked first.