Refactor AP generator's composition of permissions
Before a session for a role was faked to resolve permission for that role.
Then that condition was string concat with an inline EQL for the roles,
after attempting to group the roles together.

It was terrible.

Now we have a new executor entry point: `forEdgeDB`.
`resolve()` picks the applicable policies for the session's roles first.
`forEdgeDB()` is unaware of the current roles, and instead
adds the roles as a condition when evaluating the policy.

This is so much cleaner.

---

This also changes two logical things:

# Calculated actions

Calculated actions are generated as global access allowed.
This was a miss before.
Since we had a select AP declared but no insert/delete APs,
no one was allowed access to do those actions.
Since users aren't allowed direct access to this,
we allow "anyone" (API system) now.

# anonymous user vs dev adhoc

Before, there was confusion between how we handle an anonymous API user
and an adhoc query outside of API.
Part of that was the confusion with roles/policies, which is cleaned up
now with the role condition.
This now appropriately doesn't have a role condition when one is not needed.
If a role is needed, and the user is anonymous, then it appropriately evals to false.

Finally, for devs needing access in adhoc ways, they need to now explicitly disable APs.
Which can be done on CLI for local instances or UI for adhoc queries.