Optimize role + condition unions
Looks for:
    (role {x} and X)
    or
    (role {y} and X)

And transforms to:
    (role {x, y} and X)

This is a hot path for us, since our policies are often defined by roles,
and multiple policies declare the same conditions.
So this helps merge those same conditions across all policies into a single condition, better.
Which is an optimization for EdgeDB Access Policies.

This takes the cost of
```edgeql
select TranslationProject {*}
```
from 1,450,150 -> 486,194
That is a 3x increase.