Fix custom codecs used within EdgeDB's Session
Two changes here.

First, every pre-existing connection needs to have its (stale) session codec invalidated.
There has to be at least one, because we need to make a
connection to the DB to get the UUIDs of our schema's custom scalars.
The next time the connection is used, it will re-evaluate a fresh session codec from the codec registry.

Second, when we register our custom codecs, we clear the registry first.
This prevents (non-scalar) codecs from depending on stale (scalar) codecs.
For example, the Session codec holds a reference to the UUID codec for currentUserId,
and that becomes stale after we set a new value for the UUID codec.

This IMO is the hardest, but the most robust/future-proof way to do this.
This currently doesn't have any functional change.
But if we were to change our UUIDs to be shortened, we need this before
we could use a shortened UUID for the currentUserId global.
Since UUID has a static id, we could just assign that immediately to the
registry before the first connection is made.
But if we ever added a global for a custom scalar from our schema,
this would not work.