Enhance OptionsContext to allow changing the current layer's options
This is done via BehaviorSubjects.
A BS can be passed in to have the caller give updates at will.
Or `override` can be called to give a new value to the current layer.

Part of what makes this hard is that withOptions assumes the parent
options, and just modifies what is needed.
So changes to accessor's options also need to affect all descendants.