Rename EdgedB.withOptions -> usingOptions
This IMO better reflects the functionality.
Similar to the new `using` keyword.
It is "do something else" while "using" this functionality.

While "with" is more like mutate X1 into X2.
It is the same type just with a different state.

Also adjusted the function to just forward to the options context function,
so that the type signature doesn't have to be duplicated.