Change departmentId to str
I previously assumed it could be cast in the query,
but this only works if the field name is changed.
So for app code to function the same, this has to be
the right type in schema.

This also allows leading zeros, which could be coming
with multiplication projects.

The constraint is changed to expect a positive 5-digit integer.

Also running the rewrite on insert to enforce constraints
on projects created as active.