Run project seeds one at a time / convert to TS
Only one project can be inserted at a time per query/transaction now, which would set its departmentId.
This is because the departmentId generation logic is unaware of newly used IDs within the existing transaction.
So it will try to re-use the same ID multiple times leading to constraint errors.