Fix field condition not unwrapping secured values
Looking back, the former `StatusCondition` also didn't do this.

Due to calling code, this condition always worked fine when
securing the object because the object wasn't secured.
(Which is also what declares the `canDelete` boolean.)
It is only when using a secured object as the context,
that this bug was hit.

So this did in fact work correctly when it was written: #2863
It was only after we changed the `delete` check,
which _did_ use a secured object, that this bug was exposed: #2976