Unhack handling of exceptions thrown from field resolvers
38e7850 had a bug that 75559ba avoided but only for GraphQLErrors,
which was dumb because any kind of error can be thrown from field resolvers.

38e7850 tried to fake an execution context.
But this code doesn't work
```ts
GqlExecutionContext.create(
  new GqlExecutionContext([])
)
```
This throws, which was the error we were previously avoiding.
The `GqlExecutionContext.create()`
call is expecting an `ExecutionContextHost` (the normal one).
And if you give it the GQL one it throws.
This breaks liskov sub principle and is surprising.

Now we avoid the `ExceptionFilter.catch` method, which required
the `ExecutionContext/ArgumentsHost` parameter.
Using `undefined` instead of trying to fake the object is way easier
to reason about esp here with the types being very loose.