Restart migration chain to ease db load
Certain instances got stuck on the previous 00006 because it was too big.
This represented a fundamental schema change with #3192.
So getting this in from the get-go will be easier on from-scratch instances.

I intentionally split our current schema into these four parts to not overwhelm the DB.
We'll see though.