Decouple explicit references express via ~/core/http aliases
This really only makes working with types easier.
They can be inferred now or at least imported from a single spot.

This allows them to be swapped out easier.

Note this doesn't actually decouple the app from express,
since express API is still used.
But swapping out with something that is 80% compatible will be easier.
Only the 20% of usage has to change, and all the type refs don't have to change.