Merge sibling where clauses
This is a different approach to what the previous commit undid.
This continues to allow filters to do whatever, in dev order.
If they happen to just be a where clause and there's multiple next to each
other now we merge those together implicitly.

This prevents the broken cypher that was part of the original reason
why the `filter.builder` abstraction was created.