Consider all "missing" summary percents to be 0
We don't really know if an empty cell is a valid 0 or coming from other cells not being updated.

As long as the reporting quarter is valid, we'll count these zeros as valid.
That appeared to be the root of the problem that we were trying to solve before,
plus just trying to guard against bad xlsx navigation which has been vetted at this point.