Restore 3rd party session cookies for localhost (for Chromium & FF)
Partially reverts f29286574bebd73c5ec66cd8cb74e5e991f77aff

We want to allow 3rd party cookies (aka SameSite=none && Secured).
FF & Chromium have exceptions that allow `Secured` to work with `localhost`.
This restores that support when running locally.
One example of this is https://studio.apollographql.com/sandbox
That is hosted by a 3rd party and points to our localhost.

We also want the UI to actually work.
Safari does not have this exception.
So when trying to establish a session in Safari,
the set-cookie header is ignored because it had the Secured requirement,
and we were loading it over http (localhost).

So now we allow 3rd party cookies, except for Safari/localhost combo.