Stop assuming PnP project end date is the end of the month (#3349)
I'm not sure why we were doing this in the first place. Presumably, authors would mistakenly write the first day of the month when they meant the last, and this was to help "correct" that for them.

Now we're hitting cases where the project doesn't end on the last day of the month, and we need to allow that.

Reverts 291f47f