Set custom string tags for temporal & rich text scalars
This will prevent them from being treated as "regular objects"
So when we're blindly traversing through data objects,
we know to not traverse into the internals of these objects.

I'm also marking these to skip class transformation.
Currently, we have them transformed into these scalar instances
with the class-transformer library.
In the future, we could have the GQL custom scalars themselves do this
instead, and this marking prevents the class-transformer from choking on them.