Adjust ID validator to handle arrays with the `each` option.
The validator calls the validate method once for each item in the array.
This would normally be good, but we're abusing it to dynamically resolve
the input to different IDs.
In doing this we were clobbering the array into the first item.

Now we handle this array case specifically and do the resolution once
for the entire list.