Fix default last chapter/verse in scripture input (#3363)
I'm guessing this regressed with a7187c6.
Fixed by using an actual getter/setter for the `book` in the initial `ScriptureReferenceInput` class. This allows for overriding in `ScriptureReferenceEndInput`. The getter/setter is re-defined to be a normal property on the first set. So this makes it function like it did before at runtime.

This issue was exacerbated by the fact that this failure to produce a valid scripture reference still produced a `Verse` object that happened to be invalid. This invalid `Verse` object caused an infinite loop in `mergeVerseRanges`. This was fixed here:
https://github.com/SeedCompany/libs/commit/b02ae4033c5bc573eb3d392c4e424051b3df45b8 Where runtime checks were added to prevent those invalid objects from being created, thus avoiding the infinite loop.

For some reason Gel doesn't like this getter/setter change. Typecasting to a `Simplify`ed version restores the type compat.