Ignore suspended engagements when transitioning project
This allows a suspended engagement to exist, while the rest of the project
moves through the CtP process, for example.

We will still consider Suspended to be an ongoing status, that way the project
can't complete with a suspended engagement, it has to be resolved in some way.

We will still consider these engagements when trying to move out of suspended,
so the suspended engagements will follow the project discussing reactivation/termination