Actually fix the project step properties to be permanent immediately
This is needed to determine the correct "back" actions.

This was supposed to be achieved with ad5bb79, but that commit didn't work,
because `null ?? default` still uses `default`.

This good be the source of these hard to reproduce workflow edge cases,
as they needed multiple actions within 30 mins (default duration for permanency).