Retry assigning department ID on collision
Since we pick the min number, instead of a random one, this is duplicate can happen
when transitioning multiple projects at once.

Transitioning multiple projects at once is still extremely unlikely.

This same situation can happen for Gel, but it is not accounted for here, since it happens
in a status change.
We need to handle that separately.