Refactor Neo4jError casting
Before we recreated the error in a new instance of subclass.
Now we use `Symbol.hasInstance` to define a custom function for `instanceof` checks.
This allows us to keep the same error instance but still have easy instance checks in
the codebase.
We modify in the error in place when needed to tweak info.

I'm also non-enumerating the new error info.