Declare PolicyModule as dependent on CoreMode
This was not needed before because CoreModule is global.
And AuthorizationModule is global, which imports this module.

Now it is necessary to help Nest know that Core needs to initialize before this module.

Specifically, the PolicyFactory using the GraphQL schema, via the ResourcesHost, to determine interfaces.
And the schema needs to be built (on init) before the PolicyFactory.onModuleInit is called.