Revert #3309 / 1fa42ad for EnforceChangesetEditablePipe (#3408)
I didn't realize it at the time, but because this is a global pipe, it applies as a dependency to all controllers/resolvers. Making this request scoped means all controllers/resolvers become request scoped as well.

Revert to using `GqlContextHost`/ALS and add `contextMaybe` to prevent the former try/catch need.