Convert resource plugin to JS
I'm having a hard time getting the transformer to load & transpile.
There are too many different loading variations between Nest, Jest, ts-node, ts-patch.
It was giving a cryptic error with gel:gen (ts-node).

I don't want to get stuck here because the landscape is going to change as we switch
to SWC, vitest, etc.

There's possibly a good performance gain to do this anyway, unsure, unmeasured.
The JS with jsdocs is _almost_ as good - TS does check it.