Adjust docker image for #3415
```
yarn gel:gen
```
now runs the standard ts compilation which generates the js files in dist/

Previously the gel user did not have permission to this dir,
as we only applied the owner change to src/ (old line 87).

Now we'll set the current user as "gel" and copy in all files as that owner.
This is more idiomatic for Docker.