Replace remaining direct usages of static Props/SecuredProps (#3422)
This indirection allows them to be provided in different ways, like with a transformer, or external metadata generation.

The BaseNodeProps stuff here is clunky. I don't like it, and it could be refined more. But it is not wide spread and maybe we just wait and delete it with Gel migration.