Verify logged-in at interceptor for all mutations unless the decorator overrides
Before we required that the resolver had a parameter
```ts
@LoggedInSession() session: Session
```
Which was clunky because we had to have it even if we didn't use it.

We almost always needed to pass it down, so it made sense.
Now, with the session moving to an ALS, we will rarely need it.

So this moves the verification to the interceptor to be run regardless of
handler parameters.